in-domain ETTm1 results
Pre-trained and fine-tuned on ETTm1 dataset, the fine-tuning was done on 100% of the training dataset. Fine-tuning loss is the L2 Loss. For the transformers encoder, I used positional encoding even if it wasn't mentioned in the paper, I'll later on try a version without it to see if the results match the paper better. Because rn the model is performing better than the paper for in-domain setting, which might be because of enhanced learning or a leakage of original time-series information during the reconstruction.